squash! - peter's fixes for besu connectivity use LAN IP
Instead of hardcoded localhost use the LAN IP of machine so that the
gateway container can access it too.

Why though? Because if you tell the gateway container that it should
access the besu ledger via localhost then it will try to do that through
the container's own localhost, which is different from the host machine's
localhost (where the besu ledger is actually running).

Using the actual IP address of the host machine's primary network interface
resolves the ambiguity between the two differnet localhosts.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>